attempt to fix kill_manager
