+++ environment.yml
@@ -9,7 +9,6 @@ dependencies:
   - dill
   - ipyparallel
   - mpi4py
-  - parse
   - psutil
   - pyzmq
   - structlog
