+++ requirements.txt
@@ -3,7 +3,6 @@ dill
 ipyparallel
 mpi4py
 pandas
-parse
 psutil
 pyzmq
 structlog
